Reuse as much as possible between Email and SMS OTP
